ex normal

+++ hw9-3.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 
 int main(void) {
-    int matrix1[2][3];
-    int matrix2[3][2];
+    float matrix1[2][3];
+    float matrix2[3][2];
 
     for (int i = 0; i < 6; i++) {
-        scanf("%d", &matrix1[i / 3][i % 3]);
+        scanf("%f", &matrix1[i / 3][i % 3]);
     }
     for (int i = 0; i < 6; i++) {
-        scanf("%d", &matrix2[i / 2][i % 2]);
+        scanf("%f", &matrix2[i / 2][i % 2]);
     }
     
     printf("The first matrix you entered is\n");
     for (int i = 0; i < 2; i++) {
         for (int j = 0; j < 3; j++) {
-            printf("%d ", matrix1[i][j]);
+            printf("%f ", matrix1[i][j]);
         }
         printf("\n");
     }
@@ -22,7 +22,7 @@ int main(void) {
     printf("The second matrix you entered is\n");
     for (int i = 0; i < 3; i++) {
         for (int j = 0; j < 2; j++) {
-            printf("%d ", matrix2[i][j]);
+            printf("%f ", matrix2[i][j]);
         }
         printf("\n");
     }
@@ -31,11 +31,11 @@ int main(void) {
     printf("The multiplication product of matrix A and B is\n");
     for (int i = 0; i < 2; i++) {
         for (int j = 0; j < 2; j++) {
-            int sum = 0;
+            float sum = 0;
             for (int k = 0; k < 3; k++) {
                 sum += matrix1[i][k] * matrix2[k][j];
             }
-            printf("%d ", sum);
+            printf("%.2f ", sum);
         }
         printf("\n");
     }
